Move truck availability functions to their own file
