+++ appOPHD/TruckAvailability.h
@@ -1,0 +1,6 @@
+#pragma once
+
+
+int getTruckAvailability();
+int pullTruckFromInventory();
+int pushTruckIntoInventory();
