+++ appOPHD/UI/MineOperationsWindow.cpp
@@ -8,6 +8,7 @@
 #include "../Constants/Strings.h"
 #include "../Constants/UiConstants.h"
 #include "../StructureManager.h"
+#include "../TruckAvailability.h"
 #include "../MapObjects/Structures/MineFacility.h"
 #include "../MapObjects/Structures/Warehouse.h"
